requirements.txt
absl-py==0.7.1
astor==0.7.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.3
mkl-fft==1.0.10
mkl-random==1.0.2
mock==2.0.0
networkx==2.3
numpy==1.16.2
olefile==0.46
pandas==0.24.2
pbr==5.1.3
Pillow==5.4.1
protobuf==3.7.0
pycparser==2.19
pyparsing==2.3.1
python-dateutil==2.8.0
pytorch-nlp==0.4.1
pytz==2018.9
PyWavelets==1.0.3
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
torch==1.0.1.post2
torchvision==0.2.2
tornado==6.0.1
tqdm==4.31.1
urllib3==1.25.3
Werkzeug==0.15.0
